Go through all creds in prefs until we find unexpired order (#13773)
* Fixed credentials search from prefs

+++ components/skus/browser/rs/lib/src/models.rs
@@ -124,7 +124,7 @@ pub struct OrderItem {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct CredentialSummary {
     pub order: Order,
-    pub remaining_credential_count: u32,
+    pub remaining_credential_count: usize,
     pub expires_at: Option<NaiveDateTime>,
     pub active: bool,
 }
